backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
redis==5.0.1
celery==5.3.6
elasticsearch==8.11.0
prometheus-client==0.19.0
sentry-sdk==1.38.0
python-json-logger==2.0.7
httpx==0.25.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyJWT==2.8.0
argon2-cffi==23.1.0
stripe==7.6.0
sendgrid==6.10.0
twilio==8.10.0
boto3==1.29.6
pillow==10.1.0
qrcode==7.4.2
python-magic==0.4.27
